HCK-10771: add support of all list of available directive locations for FE and RE
