+++ shared/constants/directive.js
@@ -1,0 +1,25 @@
+const DIRECTIVE_LOCATIONS = {
+	argumentDefinition: 'ARGUMENT_DEFINITION',
+	enum: 'ENUM',
+	enumValue: 'ENUM_VALUE',
+	field: 'FIELD',
+	fieldDefinition: 'FIELD_DEFINITION',
+	inputFieldDefinition: 'INPUT_FIELD_DEFINITION',
+	inputObject: 'INPUT_OBJECT',
+	interface: 'INTERFACE',
+	mutation: 'MUTATION',
+	object: 'OBJECT',
+	query: 'QUERY',
+	scalar: 'SCALAR',
+	schema: 'SCHEMA',
+	subscription: 'SUBSCRIPTION',
+	union: 'UNION',
+	fragmentDefinition: 'FRAGMENT_DEFINITION',
+	fragmentSpread: 'FRAGMENT_SPREAD',
+	inlineFragment: 'INLINE_FRAGMENT',
+	variableDefinition: 'VARIABLE_DEFINITION',
+};
+
+module.exports = {
+	DIRECTIVE_LOCATIONS,
+};
